+++ src/AlyssaCommon.cpp
@@ -443,13 +443,6 @@ unsigned char hexconv(char* _Arr) {// Lame hexadecimal string to decimal byte co
 	return ret;
 }
 
-template <typename TP> std::time_t to_time_t(TP tp) {
-	using namespace std::chrono;
-	auto sctp = time_point_cast<system_clock::duration>(tp - TP::clock::now()
-		+ system_clock::now());
-	return system_clock::to_time_t(sctp);
-}
-
 std::string LastModify(std::filesystem::path& p) {
 	std::time_t tt = to_time_t(std::filesystem::last_write_time(p));
 	std::tm* gmt = std::gmtime(&tt);